requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
numpy==1.20.3
pandas==1.2.4
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
